requirements.txt
certifi==2018.10.15
chardet==3.0.4
Django==2.1.3
djangorestframework==3.9.0
djangorestframework-simplejwt==3.3
idna==2.7
mercadopago==0.3.5
Pillow==5.3.0
pkg-resources==0.0.0
PyJWT==1.6.4
pytz==2018.7
requests==2.20.1
urllib3==1.24.1
